static_cast qsizetype QString::size() to int
QString::size() returns an int in Qt 5 but a qsizetype in Qt 6. Cast the
size to int in order to fix Clang Semantic analysis (and probably
compilation) errors like the following while parsing with Qt 6:
includepathcompletioncontext.cpp:237:39: Non-constant-expression cannot
be narrowed from type 'qsizetype' (aka 'long long') to 'int' in
initializer list [-Wc++11-narrowing]

This also fixes GCC warnings while building against Qt 6:
includepathcompletioncontext.cpp:237:48: warning: narrowing conversion of ‘line.QString::size()’ from ‘qsizetype’ {aka ‘long long int’} to ‘int’ [-Wnarrowing]
  237 |             range.setEnd({lineNumber, line.size()});
      |                                       ~~~~~~~~~^~

KF6 KTextEditor::Cursor() still takes two int parameters. The column
numbers passed by KDevelop are not going to be out of the 32-bit integer
range.

+++ plugins/clang/codecompletion/includepathcompletioncontext.cpp
@@ -234,7 +234,7 @@ public:
             }
 
             // replace the whole line
-            range.setEnd({lineNumber, line.size()});
+            range.setEnd({lineNumber, static_cast<int>(line.size())});
         } else {
             range.setEnd({lineNumber, properties.inputTo});
         }
